+++ src/settings/settings.ts
@@ -38,9 +38,6 @@ export interface ReverseEngineeringConfig extends ListConfigBase {
 /// sjasmplus
 export interface SjasmplusConfig extends AsmConfigBase {
 	// Note: In sjasmplus the 'path' can be used either for a list file or for a sld file.
-
-	// SLD files work with addresses+bank information. If this is set all addresses are turned into simple 64k addresses as with list files.
-	disableBanking: boolean;	// TODO: REMOVE
 }
 
 
@@ -565,8 +562,7 @@ export class Settings {
 				const file = {
 					path: undefined as any,
 					srcDirs: fpSrcDirs || [""],
-					excludeFiles: fpExclFiles || [],
-					disableBanking: fp.disableBanking || false
+					excludeFiles: fpExclFiles || []
 				};
 				if (fpPath)
 					file.path = Utility.getAbsFilePath(fpPath, rootFolder)